+++ matchbox-frontend/src/app/transform/transform.component.ts
@@ -1,9 +1,13 @@
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { FhirConfigService } from '../fhirConfig.service';
 import FhirClient from 'fhir-kit-client';
-import { UntypedFormControl } from '@angular/forms';
+import { FormControl } from '@angular/forms';
 import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
 import { IDroppedBlob } from '../upload/upload.component';
+import { ReplaySubject } from 'rxjs';
+import StructureMap = fhir.r4.StructureMap;
+import OperationOutcome = fhir.r4.OperationOutcome;
+import Bundle = fhir.r4.Bundle;
 
 @Component({
   selector: 'app-transform',
@@ -11,45 +15,51 @@ import { IDroppedBlob } from '../upload/upload.component';
   styleUrls: ['./transform.component.scss'],
 })
 export class TransformComponent implements OnInit {
-  structureMaps: fhir.r4.StructureMap[];
+  // The list of all structure maps, as advertised by the server
+  allStructureMaps: StructureMap[];
+
+  // The list of filtered structure maps, to display in the select box
+  public filteredStructureMaps: ReplaySubject<StructureMap[]> = new ReplaySubject<StructureMap[]>(1);
+
+  // The form control for the structure map filter
+  public structureMapFilterControl: FormControl<string> = new FormControl<string>('');
+
+  // The form control for the selected structure map
+  public structureMapControl: FormControl<string> = new FormControl<string>(null);
 
   selectedUrl: string;
 
   client: FhirClient;
-  maps: Map<String, String>;
 
   source: string;
   mimeType: string;
 
-  selectedMap: UntypedFormControl;
-
-  query = {
-    _summary: 'true',
-    _sort: 'name',
-  };
-
-  panelOpenState = false;
-
   public transformed: any;
-  errMsg: string;
-  operationOutcome: fhir.r4.OperationOutcome;
-  operationOutcomeTransformed: fhir.r4.OperationOutcome;
+  operationOutcome: OperationOutcome;
+  operationOutcomeTransformed: OperationOutcome;
 
-  constructor(
-    private data: FhirConfigService,
-    private cd: ChangeDetectorRef
-  ) {
+  constructor(data: FhirConfigService) {
     this.client = data.getFhirClient();
-    this.client.search({ resourceType: 'StructureMap', searchParams: this.query }).then((response) => {
-      this.setMaps(<fhir.r4.Bundle>response);
-      return response;
-    });
+    this.client
+      .search({
+        resourceType: 'StructureMap',
+        searchParams: {
+          _summary: 'true',
+          _sort: 'name',
+        },
+      })
+      .then((response) => {
+        this.setMaps(<Bundle>response);
+        this.filteredStructureMaps.next(this.allStructureMaps.slice());
+      });
 
-    this.selectedMap = new UntypedFormControl();
-    this.selectedMap.valueChanges.pipe(debounceTime(400), distinctUntilChanged()).subscribe((term) => {
+    this.structureMapControl.valueChanges.pipe(debounceTime(400), distinctUntilChanged()).subscribe((term) => {
       this.selectedUrl = term;
       this.transform();
     });
+
+    // Listen for changes in the filter text
+    this.structureMapFilterControl.valueChanges.subscribe(() => this.filterStructureMaps());
   }
 
   transform() {
@@ -84,8 +94,8 @@ export class TransformComponent implements OnInit {
     return JSON.stringify(this.transformed, null, 2);
   }
 
-  setMaps(response: fhir.r4.Bundle) {
-    this.structureMaps = response.entry.map((entry) => <fhir.r4.StructureMap>entry.resource);
+  setMaps(response: Bundle) {
+    this.allStructureMaps = response.entry.map((entry) => <StructureMap>entry.resource);
   }
 
   ngOnInit(): void {}
@@ -103,5 +113,21 @@ export class TransformComponent implements OnInit {
     reader.onload = () => {
       this.source = <string>reader.result;
     };
+  }
+
+  protected filterStructureMaps() {
+    if (!this.allStructureMaps) {
+      return;
+    }
+    // get the search keyword
+    let search = this.structureMapFilterControl.value;
+    if (!search) {
+      this.filteredStructureMaps.next(this.allStructureMaps.slice());
+      return;
+    }
+    search = search.toLowerCase();
+    this.filteredStructureMaps.next(
+      this.allStructureMaps.filter((structureMap) => structureMap.name.toLowerCase().indexOf(search) > -1)
+    );
   }
 }